+++ code/machine/interrupt.hh
@@ -28,7 +28,7 @@
 /// DO NOT CHANGE -- part of the machine emulation
 ///
 /// Copyright (c) 1992-1993 The Regents of the University of California.
-///               2016-2018 Docentes de la Universidad Nacional de Rosario.
+///               2016-2020 Docentes de la Universidad Nacional de Rosario.
 /// All rights reserved.  See `copyright.h` for copyright notice and
 /// limitation of liability and disclaimer of warranty provisions.
 
@@ -79,13 +79,13 @@ public:
 
     /// initialize an interrupt that will occur in the future.
     PendingInterrupt(VoidFunctionPtr func, void *param,
-                     unsigned time, IntType kind);
+                     unsigned long time, IntType kind);
 
     VoidFunctionPtr handler;  ///< The function (in the hardware device
                               ///< emulator) to call when the interrupt
                               ///< occurs.
     void *arg;  ///< The argument to the function.
-    unsigned when;  ///< When the interrupt is supposed to fire.
+    unsigned long when;  ///< When the interrupt is supposed to fire.
     IntType type;  ///< For debugging.
 };
 
@@ -140,7 +140,7 @@ public:
     ///
     /// This is called by the hardware device simulators.
     void Schedule(VoidFunctionPtr handler, void *arg,
-                  unsigned when, IntType type);
+                  unsigned long when, IntType type);
 
     /// Advance simulated time.
     void OneTick();
